listener, subcomments

+++ Risikous_IIOb12/src/main/java/com/risikous/android/SplashActivity.java
@@ -1,40 +1,60 @@
 package com.risikous.android;
 
 import android.app.Activity;
+import android.content.Context;
 import android.content.Intent;
+import android.net.ConnectivityManager;
+import android.net.NetworkInfo;
 import android.os.Bundle;
 import android.os.Handler;
+import android.widget.Toast;
 
 import com.risikous.android.xml.GetData;
+import com.risikous.android.xml.parser.ParsePublication;
 
 /**
  * Created by Excel on 22.01.2015.
  */
 
-public class SplashActivity extends Activity{
-
-    private static int SPLASH_TIME_OUT = 5000;
+public class SplashActivity extends Activity {
 
     @Override
     protected void onCreate(Bundle savedInstanceState) {
         super.onCreate(savedInstanceState);
         setContentView(R.layout.activity_splash);
 
-        this.deleteDatabase("commentDB");
-        this.deleteDatabase("subCommentDB");
-        new GetData(this);
+        if (isConnected()) {
+            this.deleteDatabase("commentDB");
+            this.deleteDatabase("subCommentDB");
+            new GetData(this, new ParsePublication.OnRequestsFinishedListener() {
+                @Override
+                public void onFinished() {
+                    Intent i = new Intent(SplashActivity.this, MainActivity.class);
+                    startActivity(i);
 
-        new Handler().postDelayed(new Runnable() {
+                    finish();
+                }
+            });
 
-            @Override
-            public void run() {
+        }else {
+            new Handler().postDelayed(new Runnable() {
+                @Override
+                public void run() {
+                    Toast.makeText(SplashActivity.this, "Keine Internetverbindung.", Toast.LENGTH_SHORT).show();
+                    Intent i = new Intent(SplashActivity.this, MainActivity.class);
+                    startActivity(i);
 
-                Intent i = new Intent(SplashActivity.this, MainActivity.class);
-                startActivity(i);
+                    finish();
+                }
+            }, 5000);
 
-                finish();
-            }
-        }, SPLASH_TIME_OUT);
+        }
+    }
+
+    public boolean isConnected() {
+        ConnectivityManager cM = (ConnectivityManager) getSystemService(Context.CONNECTIVITY_SERVICE);
+        NetworkInfo netInfo = cM.getActiveNetworkInfo();
+        return netInfo != null && netInfo.isConnectedOrConnecting();
     }
 
 }
